Port test_guids
